add amazon as a manager + add amazon tests
The manager has been added and supported in the connect util.
The manager test framework has been expanded to include and test Amazon. All tests for finished functions are passing
rm still to go
WIP on the tests for the managers for the other functions that should be apart of the interface