fixed minetype guess for odd types
to know what content-type to set files uploaded to s3 we use mimetype.guess_type. For extensions such as onnx, it cannot guess and returns None as its prediction. This cannot be handled by the boto interface and it throws an error. The value is now defauled to s3 default application/octet-stream
tests to help